binary_tree/requirements.txt
PyYAML>=6.0
pytest>=7.0
Flask>=3.0
